requirements.txt
hf-webdriver-manager>=0.2.1
selenium>=4
pybrowsers
pillow
hbutils>=0.9.0
